final touches on chapter 8

+++ chapter-08/rental-property-microservice/src/main/java/com/homeit/rental/property/service/RentalPropertyServiceJpaImpl.java
@@ -99,8 +99,13 @@ public class RentalPropertyServiceJpaImpl implements RentalPropertyService{
     }
 
     @Override
-    public Optional<RentalPropertyDTO> delete(UUID id) {
-        Optional<RentalPropertyDTO> found = get(id);
+    public Optional<RentalPropertyDTO> delete(UUID id, String authorizedUser) {
+        Optional<RentalPropertyDTO> found = get(id)
+            .filter( rp -> rp.landlordID().toString().equals(authorizedUser));
+        if(found.isEmpty()) {
+            return Optional.empty();
+        }
+
         jpaRepository.deleteById(id);
         return found;
     }
